Add working mongoDb test cases and pom fixes

+++ mongo/src/main/java/pl/edu/agh/kis/kruchy/mongo/model/PhoneNumber.java
@@ -4,16 +4,17 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 
 @Document
 public class PhoneNumber implements Serializable {
 
-    public Long getId() {
+    public BigInteger getId() {
         return id;
     }
 
     @Id
-    Long id;
+    BigInteger id;
 
     public String getNumber() {
         return number;
